Share the order fulfillment components between both panels
The seller panel had its own fulfillments card, its own pick-what-to-ship
form and its own tracking dialog, each a parallel implementation of the
operator's. They had drifted into a different product: no thumbnails or
prices on the lines, no partial-ship warning, its own spacing, and its own
words for the same things — "Parcels" for fulfillments, "Mark as shipped"
for fulfilling, "Awaiting carrier" reduced to "Pending".

The presentational half now lives in dashboard-ui and the line-item join in
dashboard-core, so both panels render the same screens. Each keeps only its
mutations, its hooks and its SDK types, and the seller adopts the operator's
wording throughout.

Three bugs went with the duplication: a menu separator drawn above the only
item, every shipping label rendered where at most one binds a parcel, and
split units shown as two lines because the seller never joined its
fulfillment items back to their line items.
